requirements.txt
fastapi
uvicorn[standard]
google-auth
google-api-python-client
httpx
python-dotenv
python-dateutil